The JSONObjectWriter class is changed.
I found a better way to write an object to JSON format. Instated of trying to hard code, I have changed the code by simply using a Gson package. It works with and has no problem, I hope. The reason I choose to use this package is, it is easy an clean way of writing an object to a JSON file. we just need to pass a Sitecollection class object to the writer instance. I hope that is okay as long as it works. Here is the link to the package to download. http://www.java2s.com/Code/Jar/g/Downloadgson222jar.htm.