Create JsonWriter.java
I found a better way to write an object to JSON format. Instated of trying to hard code, I have changed the code by simply using a Gson library. It works well and has no problem, I hope. The reason I choose to use this package is, it is easy and is a clean way of writing an object to a JSON file. we just need to pass a Study object and a filename to generate a file.  I hope that is okay as long as it works. I have uploaded the zip file.  Here is the link to the jar file to download. http://www.java2s.com/Code/Jar/g/Downloadgson222jar.htm.